Refactor: move dependabot code and stop further processing
