+++ helpers.go
@@ -2,13 +2,10 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
 	"os/exec"
 	"strings"
 	"time"
 
-	"github.com/google/go-github/v28/github"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -79,87 +76,4 @@ func getListOfVersionedRepositories(inVersion string) ([]string, error) {
 	}
 
 	return strings.Split(strings.TrimSpace(string(output)), "\n"), nil
-}
-
-func validDependabotPR(pr *github.PullRequestEvent) bool {
-	return (pr.GetAction() == "opened" || pr.GetAction() == "reopened") &&
-		pr.GetSender().GetLogin() == "dependabot[bot]"
-}
-
-func maybeVendorDependabotPR(pr *github.PullRequestEvent) error {
-	if !validDependabotPR(pr) {
-		log.Debugf("Not a valid dependabot PR (%s:%s), ignoring...\n", pr.GetSender().GetLogin(), pr.GetAction())
-		return nil
-	}
-
-	branchName := pr.GetPullRequest().GetHead().GetRef()
-	if branchName == "" {
-		return fmt.Errorf("No Head set in the PR: %v", pr)
-	}
-
-	repo := pr.GetRepo().GetName()
-	if repo == "" {
-		return fmt.Errorf("No repository name found in PR: %v", pr)
-	}
-
-	tmpdir, err := ioutil.TempDir("", repo)
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(tmpdir)
-
-	cmd := exec.Command("git", "clone", "--single-branch", "--branch", branchName, "git@github.com:mendersoftware/"+repo+".git")
-	cmd.Dir = tmpdir
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmdDir := tmpdir + "/" + repo
-
-	cmd = exec.Command("go", "mod", "tidy", "-v")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmd = exec.Command("go", "mod", "vendor", "-v")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmd = exec.Command("go", "mod", "verify")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmd = exec.Command("git", "add", ".")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmd = exec.Command("git", "commit", "--amend", "--no-edit")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-
-	cmd = exec.Command("git", "push", "--force")
-	cmd.Dir = cmdDir
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%v returned error: %s: %s", cmd.Args, out, err.Error())
-	}
-	log.Info("Successfully vendored Dependabot PR")
-
-	return nil
-
 }
